+++ CHANGELOG.md
@@ -9,8 +9,11 @@ None
 
 ## [3.5.0] - 2020-03-29
 ### Added
-- Download of group artifacts by using a regex
+- Download of specific artifacts using a regex
 - Instructions added to README how to use this new feature
+
+## Improved
+- Section about how to clone an old Nexus3 repository
 
 ### Fixed
 - Three code smells that were reported by SonarCloud